LAU-456 Remove old feature toggles (#71)
* LAU-456 Remove cookie manager V1 feature toggle

* LAU-456 Remove Flatpickr preload feature toggle

* LAU-456 Fix stylelint module error
