+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lau-frontend",
-  "version": "1.0.12",
+  "version": "1.0.13",
   "private": true,
   "engines": {
     "node": ">=14.16.0"
@@ -31,8 +31,7 @@
     "stub:lau-idam-backend": "NODE_PATH=. node src/test/service-stubs/lau-idam-backend.js"
   },
   "dependencies": {
-    "@hmcts/cookie-manager": "^0.0.5",
-    "@hmcts/cookie-manager-v1": "npm:@hmcts/cookie-manager@^1.0.0",
+    "@hmcts/cookie-manager": "^1.0.0",
     "@hmcts/info-provider": "^1.0.0",
     "@hmcts/nodejs-healthcheck": "^1.7.0",
     "@hmcts/nodejs-logging": "^3.0.4",
@@ -77,6 +76,7 @@
     "moment": "^2.29.4",
     "node-fetch": "^2.6.7",
     "nunjucks": "^3.2.2",
+    "postcss": "^8.4.14",
     "require-directory": "^2.1.1",
     "serve-favicon": "^2.5.0",
     "ssl-config": "^1.0.0",
@@ -127,8 +127,8 @@
     "sinon-chai": "^3.4.0",
     "sonar-scanner": "^3.1.0",
     "style-loader": "^3.2.1",
-    "stylelint": "^14.3.0",
-    "stylelint-config-standard-scss": "^3.0.0",
+    "stylelint": "^14.9.1",
+    "stylelint-config-standard-scss": "^5.0.0",
     "supertest": "^6.1.6",
     "ts-jest": "^27.0.5",
     "ts-loader": "^9.2.5",
