+++ src/main/assets/scss/main.scss
@@ -70,7 +70,3 @@ $color-white: #fff;
 .cookie-preference-success {
   display: none;
 }
-
-#use-cookie-manager-v1 {
-  display: none;
-}
