+++ src/main/index.ts
@@ -2,12 +2,7 @@
 import { initAll } from 'govuk-frontend';
 
 import './assets/scss/main.scss';
-
-import {initCookieManager} from './assets/js/cookie';
-import {initCookieManagerV1} from './assets/js/cookie-v1';
-const useCookieManagerV1 = document.getElementById('use-cookie-manager-v1').textContent === 'true';
-useCookieManagerV1 ? initCookieManagerV1() : initCookieManager();
-
+import './assets/js/cookie';
 import './assets/js/flatpickr';
 import './assets/js/home';
