+++ src/main/modules/nunjucks/index.ts
@@ -2,7 +2,6 @@ import * as path from 'path';
 import * as express from 'express';
 import * as nunjucks from 'nunjucks';
 import {numberWithCommas} from '../../util/Util';
-import {LaunchDarklyClient} from '../../components/featureToggle/LaunchDarklyClient';
 
 export class Nunjucks {
   constructor(public developmentMode: boolean) {
@@ -34,9 +33,6 @@ export class Nunjucks {
     env.addGlobal('nonce', app.locals.nonce);
 
     app.use(async (req, res, next) => {
-      const useCookieManagerV1 = await LaunchDarklyClient.instance.variation('cookie-manager-v-1');
-      env.addGlobal('useCookieManagerV1', useCookieManagerV1);
-
       res.locals.pagePath = req.path;
       next();
     });
