+++ src/main/routes/home.ts
@@ -1,7 +1,6 @@
 import {Application, Response} from 'express';
 import {AppRequest} from '../models/appRequest';
 import config from 'config';
-import {LaunchDarklyClient} from '../components/featureToggle/LaunchDarklyClient';
 
 async function homeHandler(req: AppRequest, res: Response) {
   const caseFormState = req.session?.caseFormState || {};
@@ -11,10 +10,7 @@ async function homeHandler(req: AppRequest, res: Response) {
   const caseSearches = req.session?.caseSearches;
   const logons = req.session?.logons;
 
-  const flatpickrPreload = await LaunchDarklyClient.instance.variation('flatpickr-preload');
-
   res.render('home/template', {
-    flatpickrPreload,
     common: {
       maxRecords: Number(config.get('pagination.maxTotal')),
     },
