+++ src/main/views/common/template.njk
@@ -31,119 +31,67 @@
   <div class="loading-overlay">
     <div class='spinner'></div>
   </div>
-  {% if not useCookieManagerV1 %}
-    <div class="global-cookie-message" id="cm-cookie-banner">
-      {{ govukCookieBanner({
-        ariaLabel: "Cookies on Log and Audit",
-        messages: [
-          {
-            headingText: "Cookies on Log and Audit",
-            html: "<p>We use some essential cookies to make this service work.</p>
-                   <p>We'd also like to use analytics cookies so we can understand how you use the service and make improvements.</p>",
-            classes: "govuk-cookie-banner__decision",
-            actions: [
-            {
-              text: "Accept analytics cookies",
-              type: "button",
-              name: "cookies",
-              value: "accept",
-              attributes: {
-              'data-cm-action': 'accept',
-              'aria-disabled': 'false'
-            }
-            },
-            {
-              text: "Reject analytics cookies",
-              type: "button",
-              name: "cookies",
-              value: "reject",
-              attributes: {
-              'data-cm-action': 'reject',
-              'aria-disabled': 'false'
-            }
-            },
-            {
-              text: "View cookies",
-              href: "/cookies"
-            }
-          ]
-          },
-          {
-            html: cookieConfirmationHTML,
-            role: "alert",
-            classes: "govuk-cookie-banner__confirmation",
-            hidden: true,
-            actions: [
-            {
-              text: "Hide this message",
-              type: "button",
-              attributes: {
-              'data-cm-action': 'hide'
-            }
-            }
-          ]
-          }
-        ]
-      }) }}
-    </div>
-  {% endif %}
 {% endblock %}
 
 {% block header %}
-  {% if useCookieManagerV1 %}
-    {{ govukCookieBanner({
-      classes: 'cookie-banner',
-      ariaLabel: 'Cookies on Log and Audit',
-      hidden: true,
-      messages: [
+  {{ govukCookieBanner({
+    classes: 'cookie-banner',
+    ariaLabel: 'Cookies on Log and Audit',
+    hidden: true,
+    messages: [
+      {
+        classes: 'cookie-banner-message',
+        headingText: 'Cookies on Log and Audit',
+        html: cookieHtml,
+        actions: [
         {
-          classes: 'cookie-banner-message',
-          headingText: 'Cookies on Log and Audit',
-          html: cookieHtml,
-          actions: [
-          {
-            classes: 'cookie-banner-accept-button',
-            text: 'Accept analytics cookies',
-            type: "button"
-          },
-          {
-            classes: 'cookie-banner-reject-button',
-            text: 'Reject analytics cookies',
-            type: "button"
-          },
-          {
-            text: 'View cookies',
-            href: "/cookies"
+          classes: 'cookie-banner-accept-button',
+          text: 'Accept analytics cookies',
+          type: "button",
+          attributes: {
+            'aria-disabled': 'false'
           }
-        ]
         },
         {
-          classes: 'cookie-banner-accept-message',
-          html: cookieAcceptHtml,
-          role: "alert",
-          hidden: true,
-          actions: [
-          {
-            classes: 'cookie-banner-hide-button',
-            text: 'Hide this message'
+          classes: 'cookie-banner-reject-button',
+          text: 'Reject analytics cookies',
+          type: "button",
+          attributes: {
+            'aria-disabled': 'false'
           }
-        ]
         },
         {
-          classes: 'cookie-banner-reject-message',
-          html: cookieRejectHtml,
-          role: "alert",
-          hidden: true,
-          actions: [
-          {
-            classes: 'cookie-banner-hide-button',
-            text: 'Hide this message'
-          }
-        ]
+          text: 'View cookies',
+          href: "/cookies"
         }
       ]
-    }) }}
-  {% endif %}
+      },
+      {
+        classes: 'cookie-banner-accept-message',
+        html: cookieAcceptHtml,
+        role: "alert",
+        hidden: true,
+        actions: [
+        {
+          classes: 'cookie-banner-hide-button',
+          text: 'Hide this message'
+        }
+      ]
+      },
+      {
+        classes: 'cookie-banner-reject-message',
+        html: cookieRejectHtml,
+        role: "alert",
+        hidden: true,
+        actions: [
+        {
+          classes: 'cookie-banner-hide-button',
+          text: 'Hide this message'
+        }
+      ]
+      }
+    ]
+  }) }}
 
   {{ govukHeader({
     homepageUrl: "/",
@@ -186,6 +134,4 @@
 
   {# Load Flatpickr at end of body as this is non-critical #}
   <link rel="stylesheet" href="assets/css/flatpickr.min.css">
-
-  <div id="use-cookie-manager-v1">{{ useCookieManagerV1 }}</div>
 {% endblock %}
