+++ src/main/views/common/webpack/css-template.njk
@@ -1,6 +1,4 @@
 <link href="/<%= htmlWebpackPlugin.files.css[0].substr(htmlWebpackPlugin.files.publicPath.length) %>" rel="stylesheet">
 
-{% if flatpickrPreload %}
-  {# Preload flatpickr to ensure it's ready on rendering #}
-  <link rel="preload" href="assets/css/flatpickr.min.css" as="style">
-{% endif %}
+{# Preload flatpickr to ensure it's ready on rendering #}
+<link rel="preload" href="assets/css/flatpickr.min.css" as="style">
