+++ src/main/views/common/webpack/css.njk
@@ -1,6 +1,4 @@
 <link href="/main-dev.css" rel="stylesheet">
 
-{% if flatpickrPreload %}
-  {# Preload flatpickr to ensure it's ready on rendering #}
-  <link rel="preload" href="assets/css/flatpickr.min.css" as="style">
-{% endif %}
+{# Preload flatpickr to ensure it's ready on rendering #}
+<link rel="preload" href="assets/css/flatpickr.min.css" as="style">
